fix a nnz bug in GSPCI_ONFLY2_LIB, adding trans_gparam for basis transformation of gparam.

+++ pyglib/pyglib/gutz/ginput.py
@@ -15,21 +15,21 @@ def save_gparam(iso=1, ispin=1, ityp_list=[0], imap_list=[0], na2_list=[2],
         nval_bot_list=[0], nval_top_list=[2]):
 
     if '-n' in sys.argv:
-        max_iter = np.int32(sys.argv[sys.argv.index('-n') + 1])
+        max_iter = int(sys.argv[sys.argv.index('-n') + 1])
 
     with h5py.File('GPARAM.h5', 'w') as f:
-        f['/iso'] = np.asarray([iso], dtype=np.int32)
-        f['/ispin'] = np.asarray([ispin], dtype=np.int32)
-        f['/num_imp'] = np.asarray([len(ityp_list)], dtype=np.int32)
-        f['/ITYP_IMP'] = np.asarray(ityp_list, dtype=np.int32) + 1
-        f['/IMAP_IMP'] = np.asarray(imap_list, dtype=np.int32) + 1
-        f['/na2_imp'] = np.asarray(na2_list, dtype=np.int32)
-        f['/gimix'] = np.asarray([imix], dtype=np.int32)
+        f['/iso'] = np.asarray([iso])
+        f['/ispin'] = np.asarray([ispin])
+        f['/num_imp'] = np.asarray([len(ityp_list)])
+        f['/ITYP_IMP'] = np.asarray(ityp_list) + 1
+        f['/IMAP_IMP'] = np.asarray(imap_list) + 1
+        f['/na2_imp'] = np.asarray(na2_list)
+        f['/gimix'] = np.asarray([imix])
         f['/gamix'] = [amix]
-        f['/gnbreset'] = np.asarray([nbreset], dtype=np.int32)
-        f['/giembeddiag'] = np.asarray([iembeddiag], dtype=np.int32)
-        f['/gmaxiter'] = np.asarray([max_iter], dtype=np.int32)
-        f['/dc_mode'] = np.asarray([ldc], dtype=np.int32)
+        f['/gnbreset'] = np.asarray([nbreset])
+        f['/giembeddiag'] = np.asarray([iembeddiag])
+        f['/gmaxiter'] = np.asarray([max_iter])
+        f['/dc_mode'] = np.asarray([ldc])
 
         f['/dc_u_avg'] = u_avg_list
         f['/dc_j_avg'] = j_avg_list
@@ -44,7 +44,7 @@ def save_gparam(iso=1, ispin=1, ityp_list=[0], imap_list=[0], na2_list=[2],
         dim_hs_list = []
         for i, na2 in enumerate(na2_list):
             if sigma_list is None:
-                sigma = np.arange(1,na2**2+1,dtype=np.int32).reshape((na2,na2))
+                sigma = np.arange(1,na2**2+1).reshape((na2,na2))
             else:
                 sigma = sigma_list[i]
 
@@ -88,28 +88,28 @@ def save_gparam(iso=1, ispin=1, ityp_list=[0], imap_list=[0], na2_list=[2],
                 f['/IMPURITY_'+str(i+1)+'/nsym_odd'] = \
                         np.asarray(sp_rotations_list[i]).shape[0:1]
 
-        f['/dim_hs_imp'] = np.asarray(dim_hs_list, dtype=np.int32)
+        f['/dim_hs_imp'] = np.asarray(dim_hs_list)
 
 
 def save_gparambands(kpt_wt, nelectron, nbmax, iso=1, ispin=1,
         symop=1, symie=1, ensemble=0,
         ismear=0, delta=0.01, ne_list=None, h1e_list=None):
     with h5py.File('GPARAMBANDS.h5', 'w') as f:
-        f['/iso'] = np.asarray([iso], dtype=np.int32)
-        f['/ispin'] = np.asarray([ispin], dtype=np.int32)
-        f['/kptdim'] = np.asarray(list(kpt_wt.shape), dtype=np.int32)
-        f['/nbmax'] = np.asarray([nbmax], dtype=np.int32)
+        f['/iso'] = np.asarray([iso])
+        f['/ispin'] = np.asarray([ispin])
+        f['/kptdim'] = np.asarray(list(kpt_wt.shape))
+        f['/nbmax'] = np.asarray([nbmax])
         if ne_list is not None:
             assert ne_list.shape[0] == kpt_wt.shape[0], \
                 ' error: ne_list and kpt_wt do not match!'
-            f['/NE_LIST'] = np.asarray(ne_list, dtype=np.int32)  # shape (nk,3)
+            f['/NE_LIST'] = np.asarray(ne_list)  # shape (nk,3)
         f['/kptwt'] = kpt_wt
-        f['/ismear'] = np.asarray([ismear], dtype=np.int32)
+        f['/ismear'] = np.asarray([ismear])
         f['/ensemble'] = [ensemble]
         f['/delta'] = [delta]
         f['/nelectron'] = [nelectron]
-        f['/symnop'] = np.asarray([symop], dtype=np.int32)
-        f['/symie'] = np.asarray([symie], dtype=np.int32)
+        f['/symnop'] = np.asarray([symop])
+        f['/symie'] = np.asarray([symie])
         if h1e_list is None:
             h1e_list = [np.zeros((2, 2), dtype=np.complex)]
         for i, h1e in enumerate(h1e_list):
